llamaindex_tutorial/requirements.txt
llama-index
python-dotenv
llama-index-llms-gemini
llama-index-embeddings-gemini
